+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## v0.3.0 - 2025-02-28
+### Added
+- Interactive editing of bookmarks' titles (`--interactive`)
+- Ability to see script's version (`--version`)
+
+### Changed
+- The help message
+- Attributes `FOLDED` and `LAST_VISIT` won't be written, unless you use `--ie` option
+
 ## v0.2.0 - 2025-02-18
 ### Added
 - Parsing of Firefox `.json` bookmark backups